Avoid too much lines in utils.py
